qc: label_vertebrae: improve text display (move script-specific hook with script) (see #1271)
